Adds error message for food id's that don't exist

+++ routes/api/v1/foods.js
@@ -13,13 +13,27 @@ Model.knex(database)
 
 router.get('/', async (req,res) => {
   const foods = await Food.query()
-  res.send(foods)
+  if (foods) {
+    res.send(foods)
+  }
+  else {
+    res.status(404).json({
+      error: "No foods here yet!"
+    })
+  }
 })
 
 router.get('/:id', async (req,res) => {
   // const food = await Food.query().findById(req.params.id).joinRelation(Meal)
   const food = await Food.query().findById(req.params.id)
-  res.send(food)
+  if (food) {
+    res.send(food)
+  }
+  else {
+    res.status(404).json({
+      error: "No food exists with that ID"
+    })
+  }
 })
 
 router.post('/', async (req,res) => {
@@ -31,14 +45,28 @@ router.post('/', async (req,res) => {
 router.patch('/:id', async (req,res) => {
   const editedFood = await Food.query()
     .patchAndFetchById(req.params.id, req.query)
-  res.send(editedFood)
+  if (editedFood) {
+    res.send(editedFood)
+  }
+  else {
+    res.status(404).json({
+      error: "No food exists with that ID"
+    })
+  }
 })
 
 router.delete('/:id', async (req,res) => {
   const deletedFood = await Food.query()
     .findById(req.params.id)
-  await Food.query().findById(req.params.id).delete()
-  res.send(`${deletedFood.name} has been deleted.`)
+  if (deletedFood) {
+    await Food.query().findById(req.params.id).delete()
+    res.send(`${deletedFood.name} has been deleted.`)
+  }
+  else {
+    res.status(404).json({
+      error: "No food exists with that ID"
+    })
+  }
 })
